chore: move findUserView to /components
