+++ src/components/sideBar/sideBar.js
@@ -2,13 +2,13 @@ import { React } from 'react'
 import classes from './sideBar.module.css';
 import Settings from '../Settings/Settings'
 import ChatsBar from '../chatSidebar/chatSidebar';
-import Contacts from '../contactsSidebar/contactsSidebar';
+import FindUserView from '../../pages/findUserView/findUserView';
 
 const sideBar = (props) => (
     <div className={classes.sideBar}>
         {props.openStatus.isSettingsOpen ? <Settings></Settings> : null}
         {props.openStatus.isChatsOpen ? <ChatsBar></ChatsBar> : null}
-        {props.openStatus.isContactsOpen ? <Contacts></Contacts> : null}
+        {props.openStatus.isContactsOpen ? <FindUserView></FindUserView> : null}
     </div>
 );
